renamed organizer -> global_model
