+++ src/linkage/global_model/__init__.py
@@ -1,0 +1,1 @@
+from linkage.global_model.global_model import GlobalModel
