--- src/linkage/organizer/__init__.py
@@ -1,1 +1,0 @@
-from linkage.organizer.global_model import GlobalModel
